# authd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NAS multi-protocol unified authentication and file management service.\nRegister or login to obtain a JWT token, then use it to access protected endpoints.",
        "title": "NAS Authd API",
        "contact": {
            "name": "NAS Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return file list for the given path. Defaults to user's home directory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List directory contents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path (default: /data/\u003cusername\u003e)",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Directory listing",
                        "schema": {
                            "$ref": "#/definitions/handler.ListFilesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Path not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently remove the specified file or directory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete a file or directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Path to delete",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "$ref": "#/definitions/handler.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Path required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stream file content as binary data.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Download a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File path to download",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Path required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/mkdir": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new directory under the user's data path.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Create a directory",
                "parameters": [
                    {
                        "description": "Directory path",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MkdirRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Directory created",
                        "schema": {
                            "$ref": "#/definitions/handler.OKPathResponse"
                        }
                    },
                    "400": {
                        "description": "Path required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/move": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move a file or directory from one path to another.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Move or rename a file/directory",
                "parameters": [
                    {
                        "description": "Source and destination paths",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MoveFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moved",
                        "schema": {
                            "$ref": "#/definitions/handler.OKResponse"
                        }
                    },
                    "400": {
                        "description": "From and to required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a file via multipart form. Target directory defaults to user's home.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target directory (default: /data/\u003cusername\u003e)",
                        "name": "path",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload successful",
                        "schema": {
                            "$ref": "#/definitions/handler.OKPathResponse"
                        }
                    },
                    "400": {
                        "description": "File field required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/share/permission": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grant or revoke POSIX ACL access for another user.\nActions: \"readonly\" (r-x), \"readwrite\" (rwx), \"remove\" (revoke all).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "share"
                ],
                "summary": "Set file share permission",
                "parameters": [
                    {
                        "description": "Permission details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SetPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission set",
                        "schema": {
                            "$ref": "#/definitions/handler.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error description"
                }
            }
        },
        "handler.ListFilesResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.FileInfo"
                    }
                },
                "path": {
                    "type": "string",
                    "example": "/data/alice"
                }
            }
        },
...
        "handler.MkdirRequest": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "path": {
                    "type": "string",
                    "example": "/data/alice/photos"
                }
            }
        },
        "handler.MoveFileRequest": {
            "type": "object",
            "required": [
                "from",
                "to"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "/data/alice/old.txt"
                },
                "to": {
                    "type": "string",
                    "example": "/data/alice/new.txt"
                }
            }
        },
        "handler.OKPathResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "path": {
                    "type": "string",
                    "example": "/data/alice/photos"
                }
            }
        },
        "handler.OKResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "handler.SetPermissionRequest": {
            "type": "object",
            "required": [
                "path",
                "target_user"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "example": "readonly"
                },
                "path": {
                    "type": "string",
                    "example": "/data/alice"
                },
                "readonly": {
                    "type": "boolean"
                },
                "target_user": {
                    "type": "string",
                    "example": "bob"
                }
            }
        },
...
        "system.FileInfo": {
            "type": "object",
            "properties": {
                "modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"file\" or \"directory\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIs...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}